Move back to .github/scripts folder and finish tests
